AT24C02 add page w/r

+++ User/periph/inc/i2c.h
@@ -21,7 +21,7 @@ public:
      * @param p pointer of data
      * @param size bytes of data
      */
-    void read(uint8_t dev_addr, uint8_t mem_addr, void *p, uint16_t size);
+    void read(uint8_t dev_addr, uint8_t mem_addr, void *p, uint32_t size);
 
     /**
      * @brief write data sized 'size' from 'p' to 'dev_addr:mem_addr'
@@ -30,7 +30,7 @@ public:
      * @param p pointer of data
      * @param size bytes of data
      */
-    void write(uint8_t dev_addr, uint8_t mem_addr, const void *p, uint16_t size);
+    void write(uint8_t dev_addr, uint8_t mem_addr, const void *p, uint32_t size);
 
 private:
     GPIO_PinType &SCL;
